revert: openai compatible type
