+++ src/renderer/src/config/models.ts
@@ -2410,7 +2410,7 @@ export function isWebSearchModel(model: Model): boolean {
     return CLAUDE_SUPPORTED_WEBSEARCH_REGEX.test(model.id)
   }
 
-  if (provider.type === 'openai') {
+  if (provider.type === 'openai-response') {
     if (
       isOpenAILLMModel(model) &&
       !isTextToImageModel(model) &&
@@ -2441,7 +2441,7 @@ export function isWebSearchModel(model: Model): boolean {
     return models.includes(model?.id)
   }
 
-  if (provider?.type === 'openai-compatible') {
+  if (provider?.type === 'openai') {
     if (GEMINI_SEARCH_MODELS.includes(model?.id) || isOpenAIWebSearch(model)) {
       return true
     }
