+++ src/renderer/src/store/index.ts
@@ -46,7 +46,7 @@ const persistedReducer = persistReducer(
   {
     key: 'cherry-studio',
     storage,
-    version: 99,
+    version: 100,
     blacklist: ['runtime', 'messages', 'messageBlocks'],
     migrate
   },
