+++ src/renderer/src/types/index.ts
@@ -162,7 +162,7 @@ export type Provider = {
   notes?: string
 }
 
-export type ProviderType = 'openai' | 'openai-compatible' | 'anthropic' | 'gemini' | 'qwenlm' | 'azure-openai'
+export type ProviderType = 'openai' | 'openai-response' | 'anthropic' | 'gemini' | 'qwenlm' | 'azure-openai'
 
 export type ModelType = 'text' | 'vision' | 'embedding' | 'reasoning' | 'function_calling' | 'web_search'
 
@@ -462,7 +462,7 @@ export type WebSearchResults =
 export enum WebSearchSource {
   WEBSEARCH = 'websearch',
   OPENAI = 'openai',
-  OPENAI_COMPATIBLE = 'openai-compatible',
+  OPENAI_RESPONSE = 'openai-response',
   OPENROUTER = 'openrouter',
   ANTHROPIC = 'anthropic',
   GEMINI = 'gemini',
